third_party/blink/web_tests/virtual/css-relative-currentcolor-disabled/external/wpt/css/css-properties-values-api/registered-property-computation-expected.txt

This is a testharness.js-based test.
Found 1 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <color> values are computed correctly [color(from currentcolor srgb b g r)]
  assert_equals: expected "color(from currentcolor srgb b g r)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
